requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
PyJWT==2.8.0
reportlab>=4.0.0
requests>=2.31.0
